create controller & service

+++ src/main/java/com/codeunity/uno/repositary/VaccinationRecordRepository.java
@@ -4,6 +4,6 @@ import com.codeunity.uno.model.VaccinationRecordEntity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VaccinationRecordRepository extends JpaRepository<VaccinationRecordEntity,Long> {
+public interface VaccinationRecordRepository extends JpaRepository<VaccinationRecordEntity,Integer> {
     
 }
